fix tools

+++ tools/database/restore.sh
@@ -9,7 +9,9 @@ mongorestore --drop
 
 ## fetch static files
 wget -O upload.idx https://academy.abrikos.pro/uploads/
-for file in `cat upload.idx | grep -o '<a .*href=.*>' | sed -e 's/<a /\n<a /g' | sed -e 's/<a .*href=['"'"'"]//' -e 's/["'"'"'].*$//' -e '/^$/ d'`
+for file in `cat upload.idx | grep -o '<a .*href=.*>' | sed -e 's/<a /\n<a /g' | sed -e 's/<a .*href=['"'"'"]//' -e 's/["'"'"'].*$//' -e '/^$/ d'`;
 do
-wget -nc -O "uploads/${file}" "https://academy.abrikos.pro/uploads/${file}"
+if [ ! -f "uploads/${file}" ]; then
+    wget -nc -O "uploads/${file}" "https://academy.abrikos.pro/uploads/${file}"
+fi
 done
